pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "audio-llm-test"
version = "0.1.0"
description = "Audio LLM test platform for evaluating speech understanding in automotive cabin conditions"
requires-python = ">=3.11"
dependencies = [
    # Audio processing
    "numpy>=1.24",
    "scipy>=1.11",
    "soundfile>=0.12",
    # Web framework
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "websockets>=12.0",
    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    # Task queue
    "arq>=0.26",
    "redis>=5.0",
    # Config
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    # LLM APIs
    "openai>=1.30",
    "google-generativeai>=0.5",
    "anthropic>=0.25",
    "httpx>=0.27",
    # Statistics & reporting
    "pandas>=2.0",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    # Utilities
    "python-multipart>=0.0.9",
    "Levenshtein>=0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "httpx>=0.27",  # for TestClient
]
whisper = [
    "openai-whisper>=20231117",
]
tts-free = [
    "edge-tts>=6.1",
    "gTTS>=2.5",
    "piper-tts>=1.2",
    "pyttsx3>=2.90",
]
tts-ml = [
    "TTS>=0.22",          # Coqui TTS
    "suno-bark>=0.1",
]
tts-google = [
    "google-cloud-texttospeech>=2.16",
]
tts-all = [
    "audio-llm-test[tts-free,tts-ml,tts-google]",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
